requirements.txt
# AI Ethics Auditor Sentinel - Dependencies
# Core machine learning and data science libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0

# Ethics and fairness specific libraries
fairlearn>=0.9.0
aif360>=0.5.0
scikit-fairness>=0.3.0

# Natural Language Processing
nltk>=3.8
spacy>=3.6.0
textblob>=0.17.1

# Data visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
jinja2>=3.1.0

# Configuration management
pyyaml>=6.0
toml>=0.10.2
click>=8.1.0

# Testing and quality assurance
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.4.0

# Logging and monitoring
loguru>=0.7.0
wandb>=0.15.0

# Web framework for API (optional)
fastapi>=0.100.0
uvicorn>=0.22.0

# Security analysis
bandit>=1.7.5
safety>=2.3.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Utilities
tqdm>=4.65.0
rich>=13.4.0
typer>=0.9.0
python-dotenv>=1.0.0